Added check-in prevention when a user has a record checked out for evaluation or submission

+++ src/user-check-out/user-check-out.service.ts
@@ -5,6 +5,7 @@ import { UserCheckOutDTO } from '../dtos/user-check-out.dto';
 import { UserCheckOutRepository } from './user-check-out.repository';
 import { EaseyException } from '@us-epa-camd/easey-common/exceptions';
 import { UserCheckOutMap } from '../maps/user-check-out.map';
+import { getManager } from 'typeorm';
 
 @Injectable()
 export class UserCheckOutService {
@@ -68,7 +69,39 @@ export class UserCheckOutService {
     return this.map.one(record);
   }
 
+  returnManager = () => {
+    return getManager();
+  };
+
   async checkInConfiguration(monPlanId: string): Promise<Boolean> {
+    const evalRecordsInProgress = await this.returnManager().query(
+      `SELECT * FROM CAMDECMPSAUX.evaluation_set es
+      JOIN CAMDECMPSAUX.evaluation_queue eq USING(evaluation_set_id)
+      WHERE mon_plan_id = $1 AND status_cd NOT IN ('COMPLETE', 'ERROR');
+      `,
+      [monPlanId],
+    );
+
+    const submissionRecordsInProgress = await this.returnManager().query(
+      `SELECT * FROM CAMDECMPSAUX.submission_set
+       WHERE mon_plan_id = $1 AND status_cd NOT IN ('COMPLETE', 'ERROR');
+      `,
+      [monPlanId],
+    );
+
+    if (
+      (evalRecordsInProgress && evalRecordsInProgress.length > 0) ||
+      (submissionRecordsInProgress && submissionRecordsInProgress.length > 0)
+    ) {
+      throw new EaseyException(
+        new Error(
+          'Record can not be checked in. It is currently being evaluated or submitted.',
+        ),
+        HttpStatus.NOT_FOUND,
+        { monPlanId: monPlanId },
+      );
+    }
+
     const result = await this.repository.delete({ monPlanId });
     return result.affected !== 0;
   }
